Rename objc.rs to nsobject.rs
Visual Studio Code did not show any language information from `rust-analyzer` when authoring the initial commit if the file was named `nsobject.rs` (which was reproducible behavior), so it was named `objc.rs` as it provided an interface into `libobjc.dylib`. The behavior no longer reproduces so rename the file to be more clear about its contents.