Force Emit __objc_imageinfo when Sending Messages
The presence of this section prevents the optimizer from removing reads through the selector symbol, and the linker accepts multiple definitions of this section. Emitting the section with each message send ensures it exists in each object file that has a selector definition, which enables the optimizer and linker to correctly unique and preserve selector use in code. This fixes #1, enabling multiple codegen units for `objc4` and its downstream dependencies.