Rename with_retained() to with_retain()
The method name `with_retained()` was derived from the ARC bridged cast qualitifer `__bridge_retained`. But, the naming is not clear in my opinion--it can read as if the method will box a retained object.

Rename it to `with_retain()` to better clarify the `Box<T>` will retain its object.