Manage Returned Object Memory from msg_send! Macro
When the return type is `id`, require an operation (`box_retain`, `box_transfer`, `claim`) and the nullability expectation (`nonnull`, `nullable`) as part of the return type C-style cast expression. This enables `msg_send!` to retain the object, if necessary, and ensure the object is released, if necessary.

This also provides the macro with enough information to implement #3.