[corefoundation][1/4] Relax Smart Pointer Type Bound
This is the first of four steps in factoring the reference counting infrastructure into a standalone crate so it can be used to implement bindings for all reference counting APIs.

The `Object` bound does not add value. While any API that requires the pointer-is-a-CFTypeRef logical contract to be honored is unsafe, the `Object` bound does not add any safety guards.

Replace the `Object` bound with `ForeignFunctionInterface`, which provides both the raw pointer type. The next commit in this stack will move the retain/release implementation into `ForeignFunctionInterface`.