[corefoundation][2/4] Refactor Retain/Release Implementation
This is the second of four steps in factoring the reference counting infrastructure into a standalone crate so it can be used to implement bindings for all reference counting APIs.

The upcoming `retain-release` will need to support arbitrary retain and release functions. Refactor the implementation here to support this by:

1. Adding a `from_get_rule` associated function to `ForeignFunctionInterface` to perform the retain of an unowned object instance. This is a better overall factoring, which enables removing a constructor by the same name from `Arc`.
2. Adding a `release` associated function to `ForeignFunctionInterface` to enable any object type to release itself.

While we're here, rename the constructor functions to match the "conversion" naming convention, which I think is a better fit for instantiating a bindings type from a raw pointer.