[corefoundation][3/4] Remove Core Foundation References from Retain/Release Infrastructure
This is the third of four steps in factoring the reference counting infrastructure into a standalone crate so it can be used to implement bindings for all reference counting APIs.

In preparation for the next commit in this series, remove all references to "Core Foundation" from the `ForeignFunctionInterface` trait and the `rc` module so they can be moved with only minor modifications into the new crate.