[retain-release][4/4] Refactor the Retain/Release Infrastructure into new Crate
This is the last of four steps in factoring the reference counting infrastructure into a standalone crate so it can be used to implement bindings for all reference counting APIs.

Move the `ForeignFunctionInterface` trait and the `rc` module out of `corefoundation` and into a new crate, `retain-release`. Update crate and README documentation as necessary, and export the moved items from the `corefoundation` create through the previous paths.