[workspace] Re-enable invalid_value Lint
This lint was issued at only one site (`lib/darwin/src/posix/unistd.rs`), which can be fixed with a minor refactor.

For completeness, there is a test for the case where the `name` argument to `confstr(3)` is invalid. Instead of transmuting an invalid value into an invalid `enum` variant (and thus introducing undefined behavior), simply factor the `confstr(3)` binding into a private function with a `name` argument of type `i32`.