[workspace] Re-enable keyword_idents_2024 Lint
This lint was issued at only one site (`lib/darwin/src/_sys/sys/stat.rs`), which can be fixed with a minor change.

`-sys` crates should match the C API as closely as possible so adding the `st_` prefix to all of the struct's fields also resolves the `-sys` anti-pattern.